Update README.md
This should be one of the last pushes we make. It's been a long 2 years working on this project with many late nights and steep learning curves but we are very proud of the end result we came too. We are so grateful to have received this opportunity and hope that one day this code can be put to use or expanded upon. 
All the best - Tim